% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeRecall.R
\name{getAccFree}
\alias{getAccFree}
\title{Obtain accuracy serial position function for a set of trials}
\usage{
getAccFree(indat, ll, nTrials = NULL, otherVars = NULL)
}
\arguments{
\item{indat}{data frame containing recalls.}

\item{ll}{List length (number).}

\item{nTrials}{Number of trials to use in denominator; if NULL, this is worked out from unique(trial_id).}

\item{otherVars}{A vector of strings specifying other variables from the data frame (e.g., condition labels) to carry in to output. NOT CURRENTLY USED.}
}
\value{
A data frame containing the following:
\describe{
\item{serpos}{serial position (as integer)}
\item{ncor}{numer of items recalled}
\item{prob}{probability of recall of each item (nTrials as denominator)}
}
}
\description{
Obtain accuracy serial position function for a set of trials
}
\examples{
(requires dplyr and magrittr):
freerec \%>\% filter(listlen==10) \%>\% group_by(ID) \%>\% do(getAccFree(.,ll=10))
}
