% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeRecall.R
\name{getFRP}
\alias{getFRP}
\title{Obtain first recall probability function for a set of trials}
\usage{
getFRP(indat, ll, otherVars = NULL)
}
\arguments{
\item{indat}{data frame containing recalls. This must contain an integer variable named `serpos`}

\item{ll}{List length (number).}

\item{otherVars}{A vector of strings specifying other variables from the data frame (e.g., condition labels) to carry in to output.}
}
\value{
A data frame containing the following:
\describe{
\item{serpos}{serial position (as integer)}
\item{serposf}{serial position (as factor)}
\item{prob}{probability of first recall}
\item{counts}{counts of recall}
}
}
\description{
Obtain first recall probability function for a set of trials
}
\examples{
(requires dplyr and magrittr):
freerec \%>\% filter(listlen==10) \%>\% group_by(ID) \%>\% do(getFRP(.,ll=10))
}
