% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeRecall.R
\name{getlagCRP}
\alias{getlagCRP}
\title{Obtain lag conditional response probability (lag-CRP) function for a set of trials}
\usage{
getlagCRP(indat, ll, doGroup = NULL, posStruct = NULL)
}
\arguments{
\item{indat}{data frame containing recalls, with variables for trial, outpos (cannot be a factor), serpos (cannot be a factor). Recalls within each trial should be ordered by output position.}

\item{ll}{List length (number).}

\item{otherVars}{A vector of strings specifying other variables from the data frame (e.g., condition labels) to carry in to output.}
}
\value{
A data frame containing the following:
\describe{
\item{lag}{lag [from -(ll-1) to (ll-1)]}
\item{prob}{probability at each lag}
\item{lagrec}{probability at each lag (same as prob; for compatability)}
\item{numer}{numerator of lag-CRP function}
\item{denom}{denominator of lag-CRP function}
}
}
\description{
Obtain lag conditional response probability (lag-CRP) function for a set of trials
}
\examples{
(requires dplyr and magrittr):
freerec \%>\% filter(listlen==10) \%>\% group_by(ID) \%>\% do(getlagCRP(.,ll=10))
}
